--- thenewboston/transfers/admin.py
@@ -1,6 +1,0 @@
-from django.contrib import admin
-
-from .models import Block, Transfer
-
-admin.site.register(Block)
-admin.site.register(Transfer)
