--- thenewboston/transfers/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class TransfersConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'thenewboston.transfers'
