--- thenewboston/transfers/models/__init__.py
@@ -1,2 +1,0 @@
-from .block import Block  # noqa: F401
-from .transfer import Transfer  # noqa: F401
