--- thenewboston/transfers/serializers/transfer.py
@@ -1,10 +1,0 @@
-from rest_framework import serializers
-
-from ..models import Transfer
-
-
-class TransferSerializer(serializers.ModelSerializer):
-
-    class Meta:
-        model = Transfer
-        fields = '__all__'
