--- thenewboston/transfers/urls.py
@@ -1,8 +1,0 @@
-from rest_framework.routers import SimpleRouter
-
-from .views.transfer import TransferViewSet
-
-router = SimpleRouter(trailing_slash=False)
-router.register('transfers', TransferViewSet)
-
-urlpatterns = router.urls
