+++ thenewboston/wallets/admin.py
@@ -1,5 +1,7 @@
 from django.contrib import admin
 
-from .models import Wallet
+from .models import Block, Wallet, Wire
 
+admin.site.register(Block)
 admin.site.register(Wallet)
+admin.site.register(Wire)
