+++ thenewboston/wallets/models/__init__.py
@@ -1,1 +1,3 @@
+from .block import Block  # noqa: F401
 from .wallet import Wallet  # noqa: F401
+from .wire import Wire  # noqa: F401
