+++ thenewboston/wallets/models/wire.py
@@ -4,20 +4,20 @@ from django.utils.translation import gettext_lazy as _
 from .block import Block
 
 
-class TransferType(models.TextChoices):
+class WireType(models.TextChoices):
     DEPOSIT = 'DEPOSIT', _('Deposit')
     WITHDRAW = 'WITHDRAW', _('Withdraw')
 
 
-class Transfer(Block):
+class Wire(Block):
     core = models.ForeignKey('cores.Core', on_delete=models.CASCADE)
     owner = models.ForeignKey('users.User', on_delete=models.CASCADE)
     created_date = models.DateTimeField(auto_now_add=True)
-    transfer_type = models.CharField(choices=TransferType.choices, max_length=8)
+    wire_type = models.CharField(choices=WireType.choices, max_length=8)
 
     def __str__(self):
         return (
-            f'Transfer ID: {self.pk} | '
+            f'Wire ID: {self.pk} | '
             f'Owner: {self.owner.username} | '
             f'Core: {self.core.ticker} | '
             f'Amount: {self.amount}'
